Renamed tenant-components dir to components + tab IDs alignment
